chore: move types to dedicated file
