+++ source-code/plugins/i18next/src/types.ts
@@ -1,0 +1,10 @@
+export type SerializedMessage = {
+	id: string
+	text: string
+} & MessageMetadata
+
+export type MessageMetadata = {
+	parentKeys?: string[]
+	fileName?: string
+	keyName?: string
+}
